% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignCl.R
\name{assignCl}
\alias{assignCl}
\title{Assign class for test data based
based on correlation}
\usage{
assignCl(ref, refCl, test, testCl = NULL, alternative = "two.sided")
}
\arguments{
\item{ref}{Reference data (cols: samples)}

\item{refCl}{Class assignment for reference data}

\item{test}{Test dataset}

\item{testCl}{Class testset}
}
\description{
Assign class for test data based
based on correlation
}
