% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxGraph.R
\name{boxGraph}
\alias{boxGraph}
\title{Plot some position preserving boxplots}
\usage{
boxGraph(
  y,
  x,
  fact = NULL,
  xSel = NA,
  xlim = NULL,
  ylim = NULL,
  ylab = NA,
  main = "",
  lines = F,
  cols = NULL,
  colsSel = NULL
)
}
\description{
Plot some position preserving boxplots
}
