% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvIntersect.R
\name{cnvOverlap}
\alias{cnvOverlap}
\title{Prepare data for CNV overlap comparison}
\usage{
cnvOverlap(pos, nm, sel, diff)
}
\arguments{
\item{pos}{list with positions of CNV segments}

\item{sel}{list with selected indices}

\item{diff}{list with differences}

\item{id}{names of cohorts}
}
\description{
Prepare data for CNV overlap comparison
}
