% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateDiff.R
\name{createGrids}
\alias{createGrids}
\title{Create grid}
\usage{
createGrids(dat, treat, nbin = 50, rate = T)
}
\arguments{
\item{dat}{data.frame with two vectors (e.g. umap/tSNE) as first two 
colums}

\item{rate}{calulate rates (divide by number of rows for the 
respective treatment)}

\item{trea}{treatment vector, same order dat rows}

\item{bin}{number of bins of the resulting grid}
}
\description{
Create grid of rates or counts from two dim data+
e.g. as output from tSNE or UMAP
}
