% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{drawSegments}
\alias{drawSegments}
\title{Draw segments for boxplots}
\usage{
drawSegments(data, y, yDelta = -0.05, xyLeg = NULL, col = NULL,
  txtBottomOff = -1, segmentDelta = 0.2, lwd = 10, legText = c(0.55,
  0.02))
}
