% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivFract.R
\name{equivFract}
\alias{equivFract}
\title{Test for euqivalence of proportions}
\usage{
equivFract(data, grp, delta = 0.1, z = 1.65)
}
\arguments{
\item{data}{data.frame with 0,1, features in rows and samples in cols}

\item{grp}{classification, 2 levels}
}
\description{
Test for euqivalence of proportions
}
