% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAssoc.R
\name{findAssoc}
\alias{findAssoc}
\title{Find associations}
\usage{
findAssoc(grp, data, test = NULL, kat = "Fisher", filename = NULL,
  contParam = c("mean", "sd"), latex = F, outdir = "../reports/")
}
\description{
find associations between a given groupin
and a set of variables
}
