% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVprofiler.R
\name{findCNVClusters}
\alias{findCNVClusters}
\title{Find Clusters in CNV data}
\usage{
findCNVClusters(data, splitChr = ";", nSpl = 3)
}
\arguments{
\item{data}{data.frame with samples in columns and segments 
in rows. Allowed values: -1, 0, 1. rownames: "1;100;400" 
corresponding to chr 1, startPos: 100, endPos: 400}
}
\description{
Find Clusters in CNV data
}
