% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHCLSurv.R
\name{findHCLSurv}
\alias{findHCLSurv}
\title{Find hierarchical cluster-based groups for survival differences.}
\usage{
findHCLSurv(data, srv, depth = 7, minGrSize = 3, p.valCutoff = 0.1,
  dist = NULL)
}
\description{
Expectes data which can be clustered as well as corresponding survival
data. Different clusterign methods are then tried on this data.
}
