% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOptCutoff.R
\name{findOptCutoff}
\alias{findOptCutoff}
\title{Find two differening groups in survival data}
\usage{
findOptCutoff(data, srv, delta = 0.1)
}
\arguments{
\item{data}{vector with value to analyze}

\item{srv}{Surv object with the corresponding data}

\item{delta}{proportion in which group splitting should 
be performed}
}
\value{
data.frame with the calculated survival values 
(p.value, median survival, group sizes, cutoffsize, 
stepsize)
}
\description{
Tests different cutoffs on a continuous variable 
and calculates log-rank tests / survival differences
}

