% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOptCutoff_Model.R
\name{findOptCutoff_Model}
\alias{findOptCutoff_Model}
\title{Find two differening groups in survival data}
\usage{
findOptCutoff_Model(data, srv, frm0 = "srv~1", frm = "srv~GRP",
  stdVar = "VAL", delta = 0.1, minGrpSize = 1, subjVar = NULL,
  dist = NULL)
}
\arguments{
\item{data}{data.frame, with variables in cols}

\item{srv}{Surv object with the corresponding data}

\item{delta}{proportion in which group splitting should 
be performed}

\item{minGrpSize}{minimum group size for cutoffs}

\item{subject}{subject IDs if multiple measurements per subjects 
are to be tested}
}
\value{
data.frame with the calculated survival values 
(p.value, median survival, group sizes, cutoffsize, 
stepsize)
}
\description{
Tests different cutoffs on a continuous variable 
and calculates log-rank tests / survival differences
}
