% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOptCutoffChisq.R
\name{findOptCutoffClass}
\alias{findOptCutoffClass}
\title{Detect the best cutoff to match a categorical classification}
\usage{
findOptCutoffClass(data, class, delta = 0.1, test = "chisq")
}
\arguments{
\item{data}{vector with the continuous variables}

\item{class}{classification (two factors) to test
against}

\item{delta}{step size for alterations of}
}
\description{
Separates a continuous variable with different cutoffs 
into two groups and performs a chisquared or Fisher's 
exact test
}

