% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.R
\name{fitM}
\alias{fitM}
\title{Fit model}
\usage{
fitM(obj, interval = "predict", ...)
}
\arguments{
\item{obj}{fitModel instance}

\item{interval}{predict or confidence}
}
\description{
Fit a model with the selected features
}
