% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedEffAnalysis.R
\name{fixedEffAnalysis}
\alias{fixedEffAnalysis}
\title{Model based analysis with fixed effects}
\usage{
fixedEffAnalysis(data, pheno, frm0 = as.formula(VAL ~ 1),
  frm = as.formula(VAL ~ GRP), type = "lm", nCores = NULL,
  complete.cases = F, padj = "BH")
}
\arguments{
\item{data}{data.frame containg a sample per column and 
one gene per row}

\item{pheno}{Pheno data.frame with columns representing variables
and rows patients. Must be matched.}

\item{frm0}{Null-model formula}

\item{frm}{Full-model formula}

\item{type}{type of analysis, can be "lm" (linear), "nb" (negativ binomial),
"bin" (logistic regression)}

\item{nCores}{number of cores to use}

\item{complete.cases}{automatically remove incomplete data (NA, Inf)}
}
\description{
Model based analysis with fixed effects
}
