% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateDiff.R
\name{getQuantiles}
\alias{getQuantiles}
\title{Get Quantiles from grid}
\usage{
getQuantiles(dat, q = c(0.025, 0.975))
}
\arguments{
\item{dat}{output from createGrids()}
}
\description{
Get Quantiles from grid
}
