% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModelClass.R
\docType{methods}
\name{initialize,fitModel-method}
\alias{initialize,fitModel-method}
\title{Constructor fitModel}
\usage{
\S4method{initialize}{fitModel}(.Object, data = data.frame,
  meta = data.frame, type = character, var = character,
  dateOfCreation = POSIXct)
}
\arguments{
\item{.Object}{fitModel object}

\item{data}{feature data frame (features in rows, samples in cols)}

\item{meta}{metadata, with covariates as columns and samples in rows}

\item{type}{lm for linear model, lr for logistic regression}

\item{var}{variable of interest from the metadata dataframe}
}
\description{
Constructor for new fitModel instance
}
