% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localCor.R
\name{localCor}
\alias{localCor}
\title{Calculate genimc local correlation}
\usage{
localCor(data, len = 3, cutLow = -0.9, cutHigh = 0.9, corMethod = "pearson")
}
\description{
Calculate genimc local correlation
}
