% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumo.R
\docType{class}
\name{matrixSUMO-class}
\alias{matrixSUMO-class}
\alias{matrixSUMO}
\title{An S4 class representing a SUMO matrix experiment}
\description{
An S4 class representing a SUMO matrix experiment
}
\section{Slots}{

\describe{
\item{\code{sumoData}}{SUMO exported data}
}}

