% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimizeMissclassificationHCL.R
\name{minimizeMissclassificationHCL}
\alias{minimizeMissclassificationHCL}
\title{Minimize missclassification by step-wise increase
of evaluated candidates}
\usage{
minimizeMissclassificationHCL(
  data,
  res,
  class,
  clst_method = c("complete", "ward.D2"),
  test = "chisq",
  by = 1,
  complete.cases = T,
  nClust = 2
)
}
\arguments{
\item{data}{data.frame containing the data to be analyzed}

\item{res}{rownames of data, ordere by importance}

\item{class}{given classification vector}

\item{clst_method}{clustering method to use for}

\item{test}{chisq or fisher}
}
\description{
Minimize missclassification by step-wise increase
of evaluated candidates
}
