% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutMatrix.R
\name{mutMatrix}
\alias{mutMatrix}
\title{Create mut matrix}
\usage{
mutMatrix(files, excl = c("Silent"))
}
\arguments{
\item{files}{maf files}

\item{excl}{Variant_Classification to exclude}
}
\description{
Create mut matrix
}
