% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutTTest.R
\name{permTTest}
\alias{permTTest}
\title{Permutation t-test}
\usage{
permTTest(data, grp, B = 10000)
}
\arguments{
\item{data}{data}

\item{grp}{grouping, exactly two levels required}

\item{B}{Number of permutations}
}
\description{
Permutation t-test (Welch)
Might crash with a C stack overflow, if this happens, 
increase the size with ulimit -s 65535
}
