% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateDiff.R
\name{plotDiff2}
\alias{plotDiff2}
\title{Plot pairwise comparisons of grids}
\usage{
plotDiff2(dat, q = q)
}
\arguments{
\item{dat}{output from createGrids()}
}
\description{
Plot pairwise comparisons of grids
}
