% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotForestMV.R
\name{plotForestMV}
\alias{plotForestMV}
\title{Create forest plot for multivariate analysis}
\usage{
plotForestMV(srv, data, subject = NULL, selection = F, title = "",
  col = c("royalblue", "darkblue", "royalblue"))
}
