% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotForestSplit.R
\name{plotForestSplit}
\alias{plotForestSplit}
\title{Create forest plot}
\usage{
plotForestSplit(
  srv,
  data,
  subject = NULL,
  title = "",
  col = c("royalblue", "darkblue", "royalblue"),
  split = NULL
)
}
\description{
Create forest plot
}
