% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotForestSubgroupParam.R
\name{plotForestSubgroupParam}
\alias{plotForestSubgroupParam}
\title{Calculated univariate analysis and creates a forest plot}
\usage{
plotForestSubgroupParam(srv, data, ref = NULL, title = "",
  dist = "weibull", col = c("royalblue", "darkblue", "royalblue"))
}
\arguments{
\item{srv}{Survival object as created by survival::Surv() function,
each observation is linked to one row of the data parameter}

\item{data}{data.frame containing all variables which will be analyzed.
The class of each column determined the type of analysis: numeric cols 
will be treated as continous variable, factor and character as factors.}

\item{title}{Plot title}

\item{col}{Color vector as expected by the forestplot() function}

\item{subject}{vector identifying independent subjects}

\item{invalCut}{Cutoff to set HR, CI and p-values to empty values
if HR exceeds the provided cutoff (e.g. if models do not converge)}

\item{removeInval}{Retain as invalid identified levels (invalCut)}
}
\description{
The function creates a forest plot for a given 
number of variables, expect a srv object and a data.frame containing 
the selected variables as columns. Univariate Cox PH models 
are fitted. A subject vector can be specified to allow for the 
analysis of multiple observations per patient (e.g. paired samples),
by using marginal model [cluster(subject)]. Errors might occur if the graphic
devices dimension is too small (foresplot() fails).
}
