% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateDiff.R
\name{quantifyDiff}
\alias{quantifyDiff}
\title{Quantify differences from pairwise comparisons}
\usage{
quantifyDiff(dat)
}
\arguments{
\item{dat}{output from createGrids()}
}
\description{
Quantify differences from pairwise comparisons
}
