% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleAdj.R
\name{sampleAdj}
\alias{sampleAdj}
\title{Sample clusters from adjacence matrix}
\usage{
sampleAdj(adj, size = 5, n = 5, maxTry = 1000, starTopo = 1)
}
\arguments{
\item{adj}{Adjazenzmatrix}

\item{size}{Size of clusters (number of links)}

\item{n}{Number of clusters}

\item{maxTry}{numbers of random tries to establish clusters}

\item{starTopo}{create star topology}
}
\description{
Desc
}
