% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similiarity.R
\name{simLoc450k}
\alias{simLoc450k}
\title{Calculate localtion sensitive similarity, 450k}
\usage{
simLoc450k(
  data,
  type = "both",
  n = 3,
  corM = "pearson",
  cutPosCor = 0.9,
  cutNegCor = -0.9
)
}
\arguments{
\item{type}{pos, neg, both}

\item{n}{number of neighboring CpGs}

\item{corM}{type of correlation}

\item{cutPosCor}{}

\item{cutNegCor}{}
}
\description{
Calculate localtion sensitive similarity, 450k
}
