% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.R
\name{testSign}
\alias{testSign}
\title{Test significance}
\usage{
testSign(
  obj,
  pAdj = "none",
  pCut = 0.05,
  frm0 = NULL,
  frm = NULL,
  singlePVar = "Pr...t..",
  ...
)
}
\arguments{
\item{obj}{fitModel object}

\item{pAdj}{p-value adjustment methods, as in p.adjust.methods}

\item{pCut}{p-value cutoff}

\item{frm0}{Null model formula}

\item{frm}{Full model formula}

\item{singlePVar}{name of the p-value column for single level/
feature tests (when using adjusted tests)}
}
\description{
Select significant features
}
