% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumo.R, R/transf.R
\name{transf}
\alias{transf}
\title{Transform expression data}
\usage{
transf(dat, dir = 1)

transf(dat, dir = 1)
}
\arguments{
\item{dir}{direction, 1: per row (default), 2: per column}

\item{data}{vector or data.frame}
}
\description{
Performs a transformation as proposed 
by Smithson & Verkuilen (2006)
}
