% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTest.R
\name{tTest}
\alias{tTest}
\title{Row-wise t-Test}
\usage{
tTest(data, grp)
}
\arguments{
\item{data}{data}

\item{grp}{grouping, exactly two levels required}

\item{B}{Number of permutations}
}
\description{
Welch's t-test and a Satterthwaite approximation.
Might crash with a C stack overflow, if this happens, 
increase the size with ulimit -s 65535
}
