% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/westfallYoung.R
\name{wy}
\alias{wy}
\title{Westfall Young p-value correction}
\usage{
wy(data, grp, B = 100)
}
\arguments{
\item{data}{data}

\item{grp}{grouping, exactly two levels required}

\item{B}{Number of permutations}
}
\description{
Performs a Westfall-Young p-value adjustment 
using a Welch's t-test and a Satterthwaite approximation.
Might crash with a C stack overflow, if this happens, 
increase the size with ulimit -s 65535
}
