% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patCharact.R
\name{createPatChar}
\alias{createPatChar}
\title{Create patient characteristics table}
\usage{
createPatChar(data, subject = NULL, subjVar = NULL, subjRef = NULL,
  na.rm = T, latex = T, outdir = "../reports/", filename = NULL)
}
