% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randEffAnalysis.R
\name{randEffAnalysis}
\alias{randEffAnalysis}
\title{Model based analysis with random effects}
\usage{
randEffAnalysis(data, pheno, frm0 = as.formula(VAL ~ 1),
  frm = as.formula(VAL ~ GRP), rand = NULL, nCores = NULL,
  reCalcREML = F)
}
